refactor: TS 5.7 rewriteRelativeImportExtensions
