+++ bin/servitsy.ts
@@ -1,0 +1,4 @@
+// > deno run --allow-net --allow-sys bin/servitsy.ts
+// > node --experimental-strip-types bin/servitsy.ts
+import { run } from '../src/cli.ts';
+run();
