+++ src/utils.ts
@@ -1,4 +1,78 @@
 import { env, versions } from 'node:process';
+import { isAbsolute, sep as dirSep } from 'node:path';
+
+export class PathMatcher {
+	#positive: Array<string | RegExp> = [];
+	#negative: Array<string | RegExp> = [];
+	#caseSensitive = true;
+
+	constructor(patterns: string[], options?: { caseSensitive: boolean }) {
+		if (typeof options?.caseSensitive === 'boolean') {
+			this.#caseSensitive = options.caseSensitive;
+		}
+		for (const input of patterns) {
+			if (typeof input !== 'string') continue;
+			const isNegative = input.startsWith('!');
+			const trimmedInput = input.slice(isNegative ? 1 : 0);
+			const pattern = trimmedInput.length > 0 ? this.#parse(trimmedInput) : null;
+			if (pattern != null) {
+				(isNegative ? this.#negative : this.#positive).push(pattern);
+			}
+		}
+	}
+
+	test(filePath: string): boolean {
+		if (this.#positive.length === 0) {
+			return false;
+		}
+		const segments = fwdSlash(filePath).split('/').filter(Boolean);
+		const matched = this.#matchSegments(segments);
+		return matched.length > 0;
+	}
+
+	#parse(input: string): string | RegExp | null {
+		if (this.#caseSensitive === false) {
+			input = input.toLowerCase();
+		}
+		if (input.includes('/') || input.includes('\\')) {
+			return null;
+		} else if (input.includes('*')) {
+			const toEscape = /([\[\]\(\)\|\^\$\.\+\?])/g;
+			const re = input.replace(toEscape, '\\$1').replace(/\*/g, '[^/]*');
+			return new RegExp(re);
+		}
+		return input;
+	}
+
+	#matchPattern(pattern: string | RegExp, value: string): boolean {
+		if (this.#caseSensitive === false) {
+			value = value.toLowerCase();
+		}
+		if (typeof pattern === 'string') {
+			return pattern === value;
+		} else if (pattern.test(value)) {
+			const matches = value.match(pattern);
+			return matches != null && matches[0] === value;
+		}
+		return false;
+	}
+
+	#matchSegments(segments: string[]): string[] {
+		return segments.filter((segment) => {
+			const positive = this.#positive.some((pattern) => this.#matchPattern(pattern, segment));
+			if (!positive) return false;
+			const negative = this.#negative.some((pattern) => this.#matchPattern(pattern, segment));
+			return positive && !negative;
+		});
+	}
+
+	data() {
+		return structuredClone({
+			positive: this.#positive,
+			negative: this.#negative,
+		});
+	}
+}
 
 export function clamp(value: number, min: number, max: number): number {
 	if (typeof value !== 'number') value = min;
@@ -27,6 +101,13 @@ export function getEnv(key: string): string {
 	return env[key] ?? '';
 }
 
+export function getLocalPath(root: string, filePath: string): string | null {
+	if (isSubpath(root, filePath)) {
+		return trimSlash(filePath.slice(root.length), { start: true, end: true });
+	}
+	return null;
+}
+
 export const getRuntime = once<'bun' | 'deno' | 'node' | 'webcontainer'>(() => {
 	if (versions.bun && (globalThis as any).Bun) return 'bun';
 	if (versions.deno && (globalThis as any).Deno) return 'deno';
@@ -36,6 +117,17 @@ export const getRuntime = once<'bun' | 'deno' | 'node' | 'webcontainer'>(() => {
 
 export function headerCase(name: string): string {
 	return name.replace(/((^|\b|_)[a-z])/g, (s) => s.toUpperCase());
+}
+
+export function intRange(start: number, end: number, limit: number = 1_000): number[] {
+	for (const [key, val] of Object.entries({ start, end, limit })) {
+		if (!Number.isSafeInteger(val)) throw new Error(`Invalid ${key} param: ${val}`);
+	}
+	const length = Math.min(Math.abs(end - start) + 1, Math.abs(limit));
+	const increment = start < end ? 1 : -1;
+	return Array(length)
+		.fill(undefined)
+		.map((_, i) => start + i * increment);
 }
 
 export function isPrivateIPv4(address?: string) {
@@ -55,15 +147,10 @@ export function isPrivateIPv4(address?: string) {
 	);
 }
 
-export function intRange(start: number, end: number, limit: number = 1_000): number[] {
-	for (const [key, val] of Object.entries({ start, end, limit })) {
-		if (!Number.isSafeInteger(val)) throw new Error(`Invalid ${key} param: ${val}`);
-	}
-	const length = Math.min(Math.abs(end - start) + 1, Math.abs(limit));
-	const increment = start < end ? 1 : -1;
-	return Array(length)
-		.fill(undefined)
-		.map((_, i) => start + i * increment);
+export function isSubpath(parent: string, filePath: string): boolean {
+	if (filePath.includes('..') || !isAbsolute(filePath)) return false;
+	parent = trimSlash(parent, { end: true });
+	return filePath === parent || filePath.startsWith(parent + dirSep);
 }
 
 /** Cache a function's result after the first call */
